Add .env file for token
