+++ 3604/2/Gemfile
@@ -1,5 +1,6 @@
 source 'https://rubygems.org'
 
+gem 'dotenv', '2.7.4'
 gem 'redis', '4.1.2'
 gem 'redis-activesupport', '5.0.7'
 gem 'reek', '5.4.0'
