+++ 3604/2/modules/geo_uploader.rb
@@ -1,0 +1,9 @@
+module GeoUploader
+  def download_last_geo(local_path)
+    File.open(local_path + 'geo.txt', 'w') { |file| file.write(geo_parse.inspect) } if geo_parse
+  end
+
+  def geo_parse
+    @geo_parse ||= payload['location']
+  end
+end
