+++ 3604/2/modules/photo_uploader.rb
@@ -1,18 +1,8 @@
 # frozen_string_literal: true
 
-require 'redis'
-
-module HelperWithMethods
-  private
-
-  def redis
-    @redis ||= Redis.new
-  end
-
-  def user_id_telegram
-    @user_id_telegram ||= payload['from']['id']
-  end
-
+module PhotoUploader
+  Dotenv.load
+  TOKEN = ENV['TOKEN']
   API_URL_TELEGRAM = "https://api.telegram.org/bot#{TOKEN}/"
   API_URL_TELEGRAM_FILE = "https://api.telegram.org/file/bot#{TOKEN}/"
   GET_FILE_URL = 'getFile?file_id='
@@ -23,19 +13,11 @@ module HelperWithMethods
     end
   end
 
-  def download_last_geo(local_path)
-    File.open(local_path + 'geo.txt', 'w') { |file| file.write(geo_parse.inspect) } if geo_parse
-  end
-
   def photo_file_path
     JSON.parse(URI.open(url_json_about_file).read, symbolize_names: true)[:result][:file_path]
   end
 
   def url_json_about_file
     @url_json_about_file ||= API_URL_TELEGRAM + GET_FILE_URL + payload['photo'].last['file_id']
-  end
-
-  def geo_parse
-    @geo_parse ||= payload['location']
   end
 end
